Factor our common testing code into model testing utils
